+++ app/src/main/java/com/fibelatti/photowidget/configure/PhotoWidgetConfigureActivity.kt
@@ -28,7 +28,7 @@ import com.fibelatti.photowidget.configure.PhotoCropActivity.Companion.outputPat
 import com.fibelatti.photowidget.model.PhotoWidgetAspectRatio
 import com.fibelatti.photowidget.model.PhotoWidgetCycleMode
 import com.fibelatti.photowidget.model.PhotoWidgetSource
-import com.fibelatti.photowidget.model.PhotoWidgetTapAction
+import com.fibelatti.photowidget.model.PhotoWidgetTapActions
 import com.fibelatti.photowidget.platform.AppTheme
 import com.fibelatti.photowidget.platform.setIdentifierCompat
 import com.fibelatti.photowidget.widget.PhotoWidgetProvider
@@ -113,6 +113,7 @@ class PhotoWidgetConfigureActivity : AppCompatActivity() {
                         onRemovedPhotoClick = viewModel::restorePhoto,
                         onCycleModePickerClick = ::showCycleModePicker,
                         onShuffleChange = viewModel::saveShuffle,
+                        onSortChange = viewModel::saveSorting,
                         onTapActionPickerClick = ::showTapActionPicker,
                         onShapeChange = viewModel::shapeSelected,
                         onCornerRadiusChange = viewModel::cornerRadiusSelected,
@@ -303,11 +304,11 @@ class PhotoWidgetConfigureActivity : AppCompatActivity() {
     }
 
     private fun showTapActionPicker(
-        tapAction: PhotoWidgetTapAction,
+        tapActions: PhotoWidgetTapActions,
     ) {
         PhotoWidgetTapActionPicker.show(
             context = this,
-            currentTapAction = tapAction,
+            currentTapActions = tapActions,
             onApplyClick = viewModel::tapActionSelected,
         )
     }