+++ app/src/main/java/com/fibelatti/photowidget/preferences/UserPreferences.kt
@@ -1,8 +1,8 @@
 package com.fibelatti.photowidget.preferences
 
+import com.fibelatti.photowidget.model.DirectorySorting
 import com.fibelatti.photowidget.model.PhotoWidgetCycleMode
 import com.fibelatti.photowidget.model.PhotoWidgetSource
-import com.fibelatti.photowidget.model.PhotoWidgetTapAction
 
 data class UserPreferences(
     val dataSaver: Boolean,
@@ -11,11 +11,11 @@ data class UserPreferences(
     val dynamicColors: Boolean,
     val defaultSource: PhotoWidgetSource,
     val defaultShuffle: Boolean,
+    val defaultDirectorySorting: DirectorySorting,
     val defaultCycleMode: PhotoWidgetCycleMode,
     val defaultShape: String,
     val defaultCornerRadius: Int,
     val defaultOpacity: Float,
     val defaultSaturation: Float,
     val defaultBrightness: Float,
-    val defaultTapAction: PhotoWidgetTapAction,
 )